requirements.txt
django==1.9.4
pyserial==3.0.1
pytz==2016.3
